Added spring security to login
The home page, sign in, and login pages are the only pages accessable without using authentication. All other pages route back to the login page for access. H2 server is still acessible without a login for simplicity. This will be changed in further versions to ensure security. At this version the only valid login credentials are "user" for username and "password" for password. In future versions all registered users will be able to acess through the login page. Aditionally, users can sign out from the home page and must re authenticate to log back in. Auth is saved in the cache for future login attempts.